+++ Tests/test_coro.cpp
@@ -42,6 +42,16 @@ TEST_CASE("Test yield/generators with YIELD_VALUE") {
         CHECK(t.returns() == "(1, 2, 3)");
     }
 
+    SECTION("exhausted generator raises stopiteration case") {
+        auto t = EmissionTest("def f():\n"
+                              "  def cr():\n"
+                              "     yield 1\n"
+                              "     yield 2\n"
+                              "  gen = cr()\n"
+                              "  return next(gen), next(gen), next(gen)\n");
+        CHECK(t.raises() == PyExc_StopIteration);
+    }
+
     SECTION("test preservation of boxed variables") {
         auto t = EmissionTest("def f():\n"
                               "  def cr():\n"
@@ -148,5 +158,69 @@ TEST_CASE("Test yield/generators with YIELD_VALUE") {
                               "             yield f'{n}!'\n"
                               "  return [x for x in tens()]\n");
         CHECK(t.returns() == "['1!', '3!', '5!', '7!', '9!']");
+    }
+    SECTION("test nested yields and unpacker.") {
+        auto t = EmissionTest("def f():\n"
+                              "  def cr():\n"
+                              "    for idx, letter in enumerate('banana'):\n"
+                              "        if letter != 'a':\n"
+                              "            yield idx\n"
+                              "        elif letter == 'n':\n"
+                              "            for x in range(5):\n"
+                              "                yield idx\n"
+                              "  return [x for x in cr()]\n");
+        CHECK(t.returns() == "[0, 2, 4]");
+    }
+    SECTION("test non-enumerable.") {
+        auto t = EmissionTest("def f():\n"
+                              "  def cr():\n"
+                              "    for x in 255:\n"
+                              "        yield x\n"
+                              "  return [x for x in cr()]\n");
+        CHECK(t.raises() == PyExc_TypeError);
+    }
+    SECTION("test yielding a tuple.") {
+        auto t = EmissionTest("def f():\n"
+                              "  def cr():\n"
+                              "     for n in range(10):\n"
+                              "         yield f'{n}!', n\n"
+                              "  return [x for x in cr()]\n");
+        CHECK(t.returns() == "[('0!', 0), ('1!', 1), ('2!', 2), ('3!', 3), ('4!', 4), ('5!', 5), ('6!', 6), ('7!', 7), ('8!', 8), ('9!', 9)]");
+    }
+    SECTION("test yielding within a try statement.") {
+        auto t = EmissionTest("def f():\n"
+                              "  def cr():\n"
+                              "     for n in range(10):\n"
+                              "         try:\n"
+                              "            yield f'{n}!', n\n"
+                              "         except ZeroDivisionError:\n"
+                              "            pass\n"
+                              "  return [x for x in cr()]\n");
+        CHECK(t.returns() == "[('0!', 0), ('1!', 1), ('2!', 2), ('3!', 3), ('4!', 4), ('5!', 5), ('6!', 6), ('7!', 7), ('8!', 8), ('9!', 9)]");
+    }
+    SECTION("test generator being closed and recreated on exhaustion"){
+        auto t = EmissionTest("def f():\n"
+                              "    def inner():\n"
+                              "        names = ['__add__', '__alloc__',]\n"
+                              "        for n in \"!@\":\n"
+                              "            for letter in names:\n"
+                              "                yield n + letter\n"
+                              "\n"
+                              "    return list(inner())");
+        CHECK(t.returns() == "['!__add__', '!__alloc__', '@__add__', '@__alloc__']");
+    }
+    SECTION("test nested generator being closed and recreated on exhaustion"){
+        auto t = EmissionTest("def f():\n"
+                              "    def inner():\n"
+                              "        names = ['__add__', '__alloc__',]\n"
+                              "        for letter in names:\n"
+                              "            yield letter\n"
+                              "\n"
+                              "    def outer():\n"
+                              "        for n in \"!@\":\n"
+                              "            for i in inner():\n"
+                              "                yield n + i\n"
+                              "    return [x for x in outer()]");
+        CHECK(t.returns() == "['!__add__', '!__alloc__', '@__add__', '@__alloc__']");
     }
 }