+++ Tests/test_generators.py
@@ -1,3 +1,6 @@
+from io import StringIO
+import pyjion
+
 
 def test_single_yield():
     def gen():
@@ -6,6 +9,7 @@ def test_single_yield():
 
     g = gen()
     assert next(g) == 1
+    assert not pyjion.info(gen).failed
 
 
 def test_double_yield():
@@ -17,6 +21,7 @@ def test_double_yield():
     g = gen()
     assert next(g) == 1
     assert next(g) == 2
+    assert not pyjion.info(gen).failed
 
 
 def test_conditional_yield():
@@ -29,6 +34,7 @@ def test_conditional_yield():
 
     g = gen()
     assert next(g) == 1
+    assert not pyjion.info(gen).failed
 
 
 def test_yields_from_iterator():
@@ -40,6 +46,7 @@ def test_yields_from_iterator():
     g = gen()
     result = list(g)
     assert result == [1, 2, 3]
+    assert not pyjion.info(gen).failed
 
 
 def test_yields_from_range_gen():
@@ -51,6 +58,7 @@ def test_yields_from_range_gen():
     for x in gen():
         result.append(x)
     assert result == ['0!', '1!', '2!', '3!', '4!', '5!', '6!', '7!', '8!', '9!']
+    assert not pyjion.info(gen).failed
 
 
 def test_yields_from_range_gen_listcomp():
@@ -60,6 +68,7 @@ def test_yields_from_range_gen_listcomp():
 
     result = [x for x in gen()]
     assert result == ['0!', '1!', '2!', '3!', '4!', '5!', '6!', '7!', '8!', '9!']
+    assert not pyjion.info(gen).failed
 
 
 def test_nested_generator():
@@ -74,6 +83,9 @@ def test_nested_generator():
                 yield f'{n}!'
 
     assert [x for x in tens()] == ['1!', '3!', '5!', '7!', '9!']
+    assert not pyjion.info(tens).failed
+    assert not pyjion.info(evens).failed
+
 
 def test_preservation_of_boxed_variables():
     def cr():
@@ -85,6 +97,8 @@ def test_preservation_of_boxed_variables():
         yield x
     gen = cr()
     assert (next(gen), next(gen), next(gen)) == ('1', '2', '3')
+    assert not pyjion.info(cr).failed
+
 
 def test_preservation_of_unboxed_variables():
     def cr():
@@ -96,12 +110,16 @@ def test_preservation_of_unboxed_variables():
         yield x
     gen = cr()
     assert (next(gen), next(gen), next(gen)) == (1, 2, 3)
+    assert not pyjion.info(cr).failed
+
 
 def test_range_generator():
     def cr():
         for n in range(10):
             yield f'{n}!'
     assert [x for x in cr()] == ['0!', '1!', '2!', '3!', '4!', '5!', '6!', '7!', '8!', '9!']
+    assert not pyjion.info(cr).failed
+
 
 def test_yield_within_branches():
     def cr():
@@ -119,6 +137,8 @@ def test_yield_within_branches():
         yield 'c'
     gen = cr()
     assert (next(gen), next(gen), next(gen)) ==  ('a', 'c', 'd')
+    assert not pyjion.info(cr).failed
+
 
 def test_yield_within_branches_for_boxable_vars():
     def cr():
@@ -136,6 +156,8 @@ def test_yield_within_branches_for_boxable_vars():
         yield 'c'
     gen = cr()
     assert tuple(gen) == ('a', 'c', 'd', 'c')
+    assert not pyjion.info(cr).failed
+
 
 def test_yield_within_branches_for_boxable_vars_as_iter():
     def cr():
@@ -151,5 +173,23 @@ def test_yield_within_branches_for_boxable_vars_as_iter():
         else:
             yield x
         yield 'c'
-    gen = cr()
     assert [x for x in cr()] == ['a', 'c', 'd', 'c']
+    assert not pyjion.info(cr).failed
+
+
+def test_nested_generator_calling_uncompiled_function():
+    def cr1():
+        with StringIO("hello") as s:
+            for letter in s:
+                yield letter
+
+    def cr2():
+        for n in "!@#%$^":
+            for i in cr1():
+                yield n + i
+
+    gen = cr2()
+
+    assert list(gen) == ['!hello', '@hello', '#hello', '%hello', '$hello', '^hello']
+    assert pyjion.info(cr1).failed
+    assert not pyjion.info(cr2).failed