+++ Tests/testing_util.h
@@ -120,12 +120,12 @@ private:
         auto prev = _PyInterpreterState_GetEvalFrameFunc(PyInterpreterState_Main());
         _PyInterpreterState_SetEvalFrameFunc(PyInterpreterState_Main(), PyJit_EvalFrame);
         auto res = PyJit_ExecuteAndCompileFrame(m_jittedcode, frame, tstate, profile);
+        CHECK(frame->f_stackdepth != -1);
+        CHECK(frame->f_lasti >= 0);
+        CHECK(frame->f_lasti * 2 < PyBytes_GET_SIZE(m_code->co_code));
         _PyInterpreterState_SetEvalFrameFunc(PyInterpreterState_Main(), prev);
         Py_DECREF(frame);
         PyGC_Collect();
-        if (m_jittedcode->j_failed){
-            printf("Failure code : %d \n", m_jittedcode->j_compile_result);
-        }
         REQUIRE(!m_jittedcode->j_failed);
         delete profile;
         return res;
@@ -134,7 +134,9 @@ private:
 public:
     explicit EmissionTest(const char *code) {
         PyErr_Clear();
+#ifdef DEBUG_VERBOSE
         printf("--- Executing Code ---\n%s \n-----------------\n", code);
+#endif
         m_code.reset(CompileCode(code));
         if (m_code.get() == nullptr) {
             FAIL("failed to compile in JIT code");
@@ -153,7 +155,9 @@ public:
         }
 
         auto repr = PyUnicode_AsUTF8(PyObject_Repr(res.get()));
+#ifdef DEBUG_VERBOSE
         printf("Returned: %s \n", repr);
+#endif
         auto tstate = PyThreadState_GET();
         REQUIRE(tstate->curexc_value == nullptr);
         REQUIRE(tstate->curexc_traceback == nullptr);
@@ -231,11 +235,10 @@ private:
         _PyInterpreterState_SetEvalFrameFunc(PyInterpreterState_Main(), PyJit_EvalFrame);
         m_jittedcode->j_profile = profile;
         auto res = PyJit_EvalFrame(tstate, frame, 0);
-
+        CHECK(frame->f_stackdepth != -1);
         _PyInterpreterState_SetEvalFrameFunc(PyInterpreterState_Main(), prev);
         Py_DECREF(frame);
         size_t collected = PyGC_Collect();
-        printf("Collected %zu values\n", collected);
         REQUIRE(!m_jittedcode->j_failed);
         return res;
     }