+++ src/pyjion/pyjit.cpp
@@ -162,7 +162,12 @@ static inline PyObject* PyJit_ExecuteJittedFrame(void* state, PyFrameObject*fram
     trace_info.cframe.previous = prev_cframe;
     tstate->cframe = &trace_info.cframe;
 
-    frame->f_stackdepth = -1;
+#ifdef DEBUG_VERBOSE
+    printf("Starting execution of frame %s at %d with state %s\n", PyUnicode_AsUTF8(frame->f_code->co_name), frame->f_lasti, frameStateAsString(frame->f_state));
+#endif
+
+    if (frame->f_state != PY_FRAME_SUSPENDED)
+        frame->f_stackdepth = -1;
     frame->f_state = PY_FRAME_EXECUTING;
 
     try {
@@ -171,8 +176,17 @@ static inline PyObject* PyJit_ExecuteJittedFrame(void* state, PyFrameObject*fram
         tstate->cframe->use_tracing = trace_info.cframe.use_tracing;
         Pyjit_LeaveRecursiveCall();
         _Py_CheckFunctionResult(tstate, nullptr, res, __func__);
+#ifdef DEBUG_VERBOSE
+        printf("Ended execution of frame %s at %d with state %s\n",
+               PyUnicode_AsUTF8(frame->f_code->co_name),
+               frame->f_lasti,
+               frameStateAsString(frame->f_state));
+#endif
         return res;
     } catch (const std::exception& e){
+#ifdef DEBUG_VERBOSE
+        printf("Caught exception on execution of frame %s\n", e.what());
+#endif
         PyErr_SetString(PyExc_RuntimeError, e.what());
         Pyjit_LeaveRecursiveCall();
         return nullptr;
@@ -540,6 +554,14 @@ static PyObject *pyjion_get_graph(PyObject *self, PyObject* func) {
     }
 
     PyjionJittedCode* jitted = PyJit_EnsureExtra(code);
+    if (jitted->j_failed){
+        PyErr_SetString(PyExc_ValueError, "Function failed to compile so it has no graph.");
+        return nullptr;
+    }
+    if (jitted->j_graph == nullptr){
+        PyErr_SetString(PyExc_ValueError, "Compiled function has no graph, graphing was not enabled when it was compiled");
+        return nullptr;
+    }
     Py_INCREF(jitted->j_graph);
     return jitted->j_graph;
 }